SNOW-1527724: Bump up awssdk and arrow build version (#733)

+++ deps/aws-sdk-cpp-1.11.283/src/aws-cpp-sdk-core/source/client/RequestCompression.cpp
@@ -86,7 +86,7 @@ iostream_outcome Aws::Client::RequestCompression::compress(std::shared_ptr<Aws::
     {
         // calculating stream size
         input->seekg(0, input->end);
-        size_t streamSize = input->tellg();
+        size_t streamSize = (size_t)input->tellg();
         input->seekg(0, input->beg);
 
         AWS_LOGSTREAM_TRACE(AWS_REQUEST_COMPRESSION_LOG_TAG, "Compressing request of " << streamSize << " bytes.");
@@ -204,7 +204,7 @@ Aws::Client::RequestCompression::uncompress(std::shared_ptr<Aws::IOStream> input
     {
         // calculating stream size
         input->seekg(0, input->end);
-        size_t streamSize = input->tellg();
+        size_t streamSize = (size_t)input->tellg();
         input->seekg(0, input->beg);
 
         AWS_LOGSTREAM_TRACE(AWS_REQUEST_COMPRESSION_LOG_TAG, "Uncompressing request of " << streamSize << " bytes.");
